File formatting and change github actions for general testing on main and SPRINT-X pull/push
