+++ .docker/requirements.txt
@@ -1,0 +1,4 @@
+Pillow
+scipy
+numpy
+matplotlib
