--- requirements.txt
@@ -1,4 +1,0 @@
-Pillow
-scipy
-numpy
-matplotlib
